Move all our sniffs to phpcs 3
- Namespaced
- New base classes (File, Sniff, Tokens...)
- Point to new abstract sniffs
- New MoodleCodeSniffer namespace (the standard name "moodle" remains
  the same).
- Removed ForbiddenNamesAsInvokedFunctionsSniff, because it has
  been removed from PHPCompatibility
- Move to use the new phpcs:xxx annotations here and there.
- Unit tests and phpdoc fixes.

Note that this is still not working ok because of PHPCompatibility:

https://github.com/PHPCompatibility/PHPCompatibility/issues/793

Because there is a problem when that standard is included as part
of another standard (we do that) and also there is some own Sniff
extending from it (we also do that).

Will try to fix that in a next commit.