New moodle.PHPUnit.TestCaseNames Sniff
Sniff in charge of verifying various things related with PHPUnit
test cases names (files and classes and namespaces). Like:
- Error: _test.php files missing any valid test case.
- Error: incorrectly names test cases (not ended with _test[case]
- Warning: test case classes not matching test case file names.
- Error: found duplicate test case names.
- Error: found test case name proposed for another test case.
- Error: test case namespace not matching expected component namespace.
- Warning: test case namespace missing, suggesting a correct one.
- Error: found duplicate proposed test case names.
- Error: found proposed test case name existing for another test case.

Practically covered with unit tests. And behat tests updated.

Also included, small modification to local_codechecker_testcase base
class that now, when an incorrect fixture file is passed, it fails
with error (previously it was exiting silently and was hard to
detect what was happening when a typo in the fixture name was used).