Downgrade code to be PHP 7.0 compliant.
We need codechecker to continue working with PHP 7.0, because
it's used by moodle-plugin-ci that supports that version.

Hence, we have to downgrade code till it's decided to raise
requirements in both products.

Not much problem, only a few code recently introduced
(using some optional return type hinting that was added
with PHP 7.1) needs to be downgraded.