Verify that sub namespaces in test match the expected location
Added one more check to the unit tests naming sniff that now
verify, for testcase clases, that they are in the expected
subdirectory, following the general namespace rules for stuff
under the "classes" directory.

Note that we still aren't checking the correctness under "classes"
and, once that's implemented, surely some code here will be moved
to the general namespaces sniff (a note about that has been added
to code).

Also, note that we aren't still checking for level2 correctness, that
must be a valid API or "local", whenever that's possible it will be
checked.

So, this only validates that the location (sub-directories) of the
file matches the sub-namespaces of the class.

Covered with tests.