Add new TestCaseCovers sniff to control @covers tags in unit tests
This new feature will be in charge of validating the presence and
correctness of PHPUnit coverage annotations/tags.

Restrictions:
- It's only applied to Moodle 4.0 and up.
- It's only applied to files under /tests directories.
- It's only applied to files named `_test.php`.
- It's only applied to classes named `_test` or `_testcase`.
- it's only applied to methods named `test_ `.

Warnings implemented:
- Warn if both @covers and @coversNothing information are missing.
- Warn when there are contradictory mixed @coversNothing (at class) and
  @covers (at method) combinations.
- Warn about redundant @coversNothing present both at class and method.

Errors implemented:
- Error when there are contradictory @coversNothing and @covers in the
  same class or method.
- Error if the @coversDefaultClass is used in method.
- Error if the @coversNothing tag has any value.
- Error if the @covers or @coversDefaultClass tags don't have any value.
- Error if the @coversDefaultClass tag doesn't start with \ (FQCN).
- Error if the @coversDefaultClass tag has :: (reserved for @covers).
- Error if the @covers tag doesn't start with \ (FQCN) or :: (method).

96.67% line coverage achieved.

Impact notes: See that, while adding @covers annotations to various
unit tests, it's also mandatory to add a description to them. That's
because the local_moodlecheck (in charge of examining phpdoc blocks)
accepts the absence of phpdoc blocks, but once they exist, the
description in mandatory. Not sure if we should relax that in
moodlecheck or no (being strict, I think it's correct).