Apply all the standard ignores to the "other" checks
When invoking phpcs we pass to it a calculated list of
paths to ignore (substring matching).

That includes all the third part libraries, some directories
like the amd/yui build ones, and the list of exclusions added
in the UI.

It works ok, so far.

But, later, when invoking the "other" checks (an ancient list
of custom checks applied to other files, surely to be removed
some day), we don't apply all those ignores commented above.

This PR just makes those calculated ignores to be also applied to
the "other checks". Covered with tests, it seems to be working ok.

Also, fixes a problem with MoodleUtil unit tests breaking CFG
for next tests.

Fixes #179.