Apply sensible default values for $CFG if not specified
In order to use the core_component functions, sensible defaults are
required.

These sensible defaults are available, but the check of `$CFG->dirroot`
when $CFG is empty lead to a php warning, causing phpcs failures.

Even if this error is fixed, the default values in $CFG are then set to
NULL after including core_component, which means that uses of it work
with null values and cause require errors.

Fixes #184