CamelCase test cases
This only changes the names of the test cases to CamelCase. Better
for consistency with the rest of the codebase, and now we can because
this code isn't Moodle code anymore.

Also, the name change has the advantage that Moodle's phpunit won't
pick the test cases anymore, as far as it requires _test.php named
files and now we are Test.php

There are still a lot of things to remove (packages...) but this is
just a first step towards making moodle-cs to have its own (PSR2 ?)
standard for own checks.