Enable test_getMoodleBranch skipped tests
In the past, when the standard was under local/codechecker
there was always a parent directory with a Moodle installation
so the moodleRoot directory was easier to infer.

Now, as a completely separated standard, that's not true anymore
so we need to mock the filesystem (with vfsStream) to emulate
that situation. Useful for whenever the standard is installed
in a moodle site or whatever, although normally we'll be passing
the moodleRoot config option near always.