Enable test_getMoodleComponent skipped tests
We are adding a "complete" mock site, which includes
all the required files (version, config-dist...), and
also the components.json file and the core_component class.

And then, a fake "local_invented" plugin to verify that
all the components are being properly loaded.

Also, rename a few variables to avoid confusion and ensure
all the config values are reset between tests.