Apply for export-ignore towards better release packages
When used as a dependency in other projects, we don't want
the release packages to include any development file. This
includes CI files, test files, docs...

This is useful to both:
- Reduce the package sizes noticeably.
- Ensure that composer, by default, (--prefer-dist) doesn't
  include them. If anybody prefers to get the complete version
  they always can use --prefer-source or clone the repository.

This is considered good practice, as far as there can be problems
with tests and other stuff when the dependencies include everything.