Pick the PHPCSExtra CommaAfterLastSniff Sniff to moodle
We need this applied ASAP, because it's causing troubles
with many arrays that are allowed, so we cannot wait for
PHPCSExtra v1.2.0 to be released.

So, with this commit, we are moving to own (copied verbatim)
Sniff. And adjusting tests and other small bits.

Note that this commit will be 100% reverted once PHPCSExtra
is released and then we'll be back to the upstream Sniff.

See https://github.com/moodlehq/moodle-cs/issues/82