Move phpcs from squizlabs to phpcsstandards
PHP_Codesniffer is being abandoned by squizlabs. See:

https://github.com/squizlabs/PHP_CodeSniffer/issues/3932 (old org)
https://github.com/PHPCSStandards/PHP_CodeSniffer/issues/109 (new org)

So it has been moved (copied, not transferred, not good, squizlabs)
to the PHPCSStandards organisation.

So this PR just moves us to start using the new organisation. Let's
see how things evolve. It would be so great if the new one gets some
support... (right now it's a 1-person project, basically).