Lock nikic/php-parser v4
It seems that both phpunit and php-code-coverage have
started to accept nikic/php-parser v5, but that's leading
to problems with PHP 7.4 runs, that we still support.

So, with this commit, we are locking nikic/php-parser v4

Once we officially make moodle-cs to only support php >= 8.0
we can remove this lock. Issue: #88