Disable package checks whenever missing a target component
All the Sniffs that play with components (package, namespaces,...)
need to skip all the checks related with null components returned
by MoodleUtil::getMoodleComponent().

That happens when moodle-cs is executed against a standalone directory
that is not within a moodle structure, so it cannot find all the
component => paths mappings or a file with that structure has been
provided using the `moodleComponentsListPath` config option.

This commit just ensures that the moodle.Commenting.Package sniff
does not run any check requiring a expected component. And covers
it with tests (some checks still work, because they don't require
any expected component).